Add dummy method to forms (quiets LastPass)
It appears LastPass looks for the use of the GET method in forms on the page to determine if it is secure or not. Since Shout never actually submits any of the forms, no method is set, which means the default of GET is used. Changes form methods to POST to quiet the warning. Should fix #648