Wrap badge and notifications in try/catch
This fixes a somewhat rare/unusual bug where if bots dumps too much output to the user (such as /msg ChanServ help commands or /msg *status help), it crashes the client and reloads. Also fixes an Android deprecation issue with notifications that also crashed the client.

As a general rule I think crashing on notifications is a bit silly so I just wrapped the whole thing in a try/catch.