Replace message parser with a new one
This one should be really hard to break/unbreakable. It should be just as complete as the old one and address #15, #128, #129 and possibly various other unexpected breakages.

It solves the problem by making it so when a parser matches, it is responsible for recalling the parser with what it allowed to be parsed further.