Fix duplicate nicks
Due to the rarity of this bug, I suspect there is some kind of race condition going on where we send a `NAMES` request, and in the meantime users have joined on the server so we get them back as a result. But then we also receive their respective `JOIN`s after the names list, so we add them to our existing list.

Well even if it's not it, it should fix the problem anyway so whatever.