Use CI caches for downloaded files instead of installed ones
This allows for a more meaningful build: if a newer version of a sub-package breaks,
builds would still pass as it uses the cached version. This uses a cache for downloaded packages instead.

I am expecting this to slow down a little bit the builds (but we are OK overall)
but be more accurate in practice.

See https://docs.npmjs.com/cli/cache#configuration and https://docs.npmjs.com/files/folders#node-modules.