Remove devDependencies badge on README
Now that Greenkeeper is taking care of our dependencies, it makes less sense to keep this one.

This also create some room for a potential future extra badge if we need it, without taking up a new line.

I would however definitely keep the production dependencies badge because it informs users on the status of what they install. It also leaves an access to David's devDependencies report if interested.