Remove nick highlighting feature to avoid highlighting non-nick words
thelounge highlights all nicks in a message and turns them into user
menu buttons.  But people sometimes use nicks that are frequently-used
words like "the" or "that".  thelounge then highlights any usage of
"the" or "that" in that channel, even when the word does not refer to
a user with that nick.  In fact, many more nicknames are regular
words, and thelounge can't possibly know if a word was used in nick or
non-nick context.  Because highlighting random words is confusing and
distracting, remove the nick highlighting feature entirely.