Use a new npmjs token to publish through Travis
npmjs has invalidated all tokens recently, so I had to create a new one.

For the record (because it will happen again one day or another, and for the life of me I can never remember how to do it), the command I ran to encrypt this was:

```
travis encrypt ****************82f5 --add deploy.api_key -r thelounge/thelounge
```

Those last 4 characters should appear in Travis logs the next time we release, to confirm the encrypted value is correct.