extend custom highlight regex
As discussed and tested on irc, this extends matching for custom highlights. Main reason was that a word defined in custom highlights followed by e.g. a colon, a comma or an exclamation mark, such as "Jay: hi!" or "Hi Jay!" wouldn't be caught by the regex.

For some tests, see https://regex101.com/r/DDGVfN/11