Make sure the editorconfig indent size applies to all file formats
I was glancing at the Vue PR, and was tired of scrolling horizontally on files with a lot of levels. I first discovered the `?ts=2` query string for GitHub, but then [read here](https://stackoverflow.com/a/33831598/1935861) that it will respect the `indent_size` for tabs, which is rather nifty.