Allow animated webp
Webp can be animated (just like gifs but better) but up until now they were run through a canvas and thus the animation was lost. Webp can contain exif/xmp etc. and thus we want to be able to remove the data. Problem here is, there is no js library out there to remove the metadata from the, only read. So instead we read the metadata from the image and then decide wether or not we want to run in through a canvas. So this means if:
- animated / metadata => static / no metadata
- animate / no metadata => animated / no metadat