Print the response time between two clients with CTCP PING
According to spec PING is used as reachability check and to measure the time
delay between two clients. Other clients such as IRSSI and HexChat give
the user an easy way to check this delay with `/ping {nick}`

As such we implement this here as well. In the background `/ping {nick}` will
do a simple `/ctcp {nick} PING {timestamp}` to which we will print the
time delay as an answer.