Exclude image/heif from metadata stripping
On some OS ImageMagick is too old to deal with HEIF. Instead of erroring
and not uploading any picture HEIF files will be uploaded as is.

When sharing HEIF pictures from iOS devices the OS takes care of
converting the picture to JPEG beforehand which then will be
successfully rotated and stripped from metadata.