Always chmod home dir on startup
Helps to prevent and retrospectively address issues like
thelounge/thelounge-deb#72

Obviously some sysadmins may want the directory to be world-readable,
but presently, that isn't safe.

This commit could be omitted or later reverted if sensitive data
protection is ensured through other means - e.g. with patches for
existing files, and an emphasis at review time to catch over-exposure in
the future ✌