Support twitch.tv nick capitalization (sent in the display-name tag)
Twitch uses this tag to send capitalization for nicknames, set by users
in Settings, and always sends the regular nick field lowercase.

A large portion of Twitch users set it, and using the wrong
capitalization to address a Twitch user looks out of place.

This sets a user's nick to the correct capitalization, when received.

And therefore also helps with autocompletion.