server: decode link-preview text using transport/document-specified encoding
This change detects the specified encoding of any document posted as a
link, and uses that encoding to display the preview text for the link.

Otherwise, without this change, the link-preview text for documents with
legacy encodings (such as Shift_JIS) doesn’t get displayed as expected,
but instead gets garbled.